+++ test/envtest/crds_envtest_test.go
@@ -30,6 +30,8 @@ import (
 // TestGatewayAPIControllersMayBeDynamicallyStarted ensures that in case of missing CRDs installation in the
 // cluster, specific controllers are not started until the CRDs are installed.
 func TestGatewayAPIControllersMayBeDynamicallyStarted(t *testing.T) {
+	t.Parallel()
+
 	scheme := Scheme(t, WithKong)
 	envcfg := Setup(t, scheme, WithInstallGatewayCRDs(false))
 
@@ -79,6 +81,8 @@ func TestGatewayAPIControllersMayBeDynamicallyStarted(t *testing.T) {
 // TestNoKongCRDsInstalledIsFatal ensures that in case of missing Kong CRDs installation, the manager Run() eventually
 // returns an error due to cache synchronisation timeout.
 func TestNoKongCRDsInstalledIsFatal(t *testing.T) {
+	t.Parallel()
+
 	scheme := Scheme(t)
 	envcfg := Setup(t, scheme, WithInstallKongCRDs(false))
 
@@ -96,6 +100,8 @@ func TestNoKongCRDsInstalledIsFatal(t *testing.T) {
 }
 
 func TestCRDValidations(t *testing.T) {
+	t.Parallel()
+
 	ctx := context.Background()
 	scheme := Scheme(t, WithKong)
 	envcfg := Setup(t, scheme)